+++ service/index/writer.go
@@ -26,9 +26,11 @@ import (
 	"golang.org/x/sync/semaphore"
 
 	"github.com/onflow/flow-go/ledger"
+	"github.com/onflow/flow-go/ledger/complete/wal"
 	"github.com/onflow/flow-go/model/flow"
 
 	"github.com/onflow/flow-archive/models/archive"
+	"github.com/onflow/flow-archive/util"
 )
 
 // Writer implements the `index.Writer` interface to write indexing data to
@@ -124,6 +126,27 @@ func (w *Writer) Payloads(height uint64, paths []ledger.Path, payloads []*ledger
 	}
 
 	return w.apply(ops...)
+}
+
+// Registers writes the given registers in a batch to database
+func (w *Writer) Registers(height uint64, registers []*wal.LeafNode) error {
+	batch := util.NewBatch(w.db)
+	writeBatch := batch.GetWriter()
+
+	for _, register := range registers {
+		op := w.lib.BatchSavePayload(height, register.Path, register.Payload)
+		err := op(writeBatch)
+		if err != nil {
+			return fmt.Errorf("could not batch write registers to database at height %v: %w", height, err)
+		}
+	}
+
+	err := writeBatch.Flush()
+	if err != nil {
+		return fmt.Errorf("could not flush write registers to database at height %v: %w", height, err)
+	}
+
+	return nil
 }
 
 // Collections indexes the collections at the given height.
